Create a product page context (#511)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Create a product page context
Remove `getBlockContext` functionality.
Subscribe to block context in product page blocks.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

This will be much easier to maintain than the previous solution.

This way we can keep everything to do with a block in one place.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code