Fix product page revalidate bug (#513)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Fix bug on Product Page when revalidating data after a mutation.

Only handle data fetching server-side for Product Page.

Pass ShopSession ID to page to subscribe to query (header).

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

This should make the page interactions work for now.

In the future we should either commit to making this page SSR or move user-dependent data to client side data fetching.

But in the mean time it's better to handle all data fetching in one place.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code