Add form template for SE accident (#593)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

See title.

![Screen Shot 2022-09-26 at 16.37.07.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/4b36fbeb-8677-450b-937d-51d67a128541/Screen%20Shot%202022-09-26%20at%2016.37.07.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code