Fix how to get a URL from Storyblok link field (#601)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Fix how to get a URL from Storyblok link field

Remove unused `useStroryblokLinkURL` hook.

Remove locale handling.

Update `LinkField` type to make story-field optional.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We don't need to fiddle with the locale since it will come directly from Storyblok.

The "story" field should be marked optional as there is no guarantee that the link field is related to an existing story.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code