Update implementation to allow going through full flow (#605)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Update implementation to allow going through full flow

Use LinkButton on CartPage.

Lowercase URL parameter.

Add temporary redirect from Payment page if After Sign flow.

Temporary comment out check if checkout is completed on confirmation page.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

It's time to get the POC together!

This is an attempt to enable us to show off as much as possible of the new experience and share it with the rest of the org.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code