Store access token in cookie (#622)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Store access token in encrypted cookie after successful signing.

Add `cookies-next` as dependency to store app. We already used it.

Pass req + res when initializing apollo to get access token from cookies server side.

Use "auth link" from official example to pick up access token from cookies client side.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We need the authorization header to make authenticated calls to the API.

We want to access the token both client and server-side so best way to store it is in a cookie.

We should probably replace the `js-cookie` dependency and just use `cookies-next`.

## Jira issue(s): [GRW-1564]

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code


[GRW-1564]: https://hedvig.atlassian.net/browse/GRW-1564?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ