Complete purchase flow before confirmation (#645)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Reset the current shop session at the end of the checkout.

Update confirmation page URL to include Shop Session ID.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

After checkout is complete we need to setup a new Shop Session (using the auth token). This will be done as soon as we navigate to the next page (confirmation).

The confirmation page needs to identify the relevant shop session by the ID in the URL instead of from cookies. This also makes the confirmation page behave more like a recipt that you can get back to.

TODO: make confirmation page a protected route when we can link a Shop Session to a specific member.

## Jira issue(s): [GRW-1577]

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code


[GRW-1577]: https://hedvig.atlassian.net/browse/GRW-1577?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ