Implement device id header (#672)
## Describe your changes

Implement the device ID setup that we have on Web Onboarding.

Bascially we create a unique UUID that we store in the browser.
Then we send this as a header with all requests to the backend.

## Justify why they are needed

I decided to link this to the Apollo Client setup to ensure that the UUID is created before we send any API requests.

I talked with Aleks about collecting things like this in a `LocalContext` -- let's iterate on exactly how it should work!

The purpose with the header is to link a `ShopSession` with a `PaymentConnection` on the backend.

## Jira issue(s): [GRW-1609]

## Checklist before requesting a review

- [x] I have performed a self-review of my code


[GRW-1609]: https://hedvig.atlassian.net/browse/GRW-1609?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ