Store access token in same cookie as web-onboarding (#679)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Store member access token the same way as in Web Onboarding project.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

This way we can share it between the two deployments - if you are logged in, in one place, you are logged in everywhere!

We will use this to support adding payments (SE) after a users has signed in the checkout-flow.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code