Implement option to share a link to a specific shop session (#961)
## Describe your changes

Add route that can be used to share a shop session using a link.

```url
/se/session/2i1oj3oij12oij12iojwio2j?next=/se/cart
```

## Justify why they are needed

This can be very helpful to share shop session with other people to debug. For example in test sessions.

We should probably build a feature that automatically generates the link.

I considered adding this feature to the middleware but I think this will be much easier to maintain.

## Jira issue(s): [GRW-1852]

## Checklist before requesting a review

- [ ] I have performed a self-review of my code


[GRW-1852]: https://hedvig.atlassian.net/browse/GRW-1852?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ