Fix layout in `OfferPresenter` (#984)
## Describe your changes

Fix layout issue in `OfferPresenter` when `CancellationForm` is visible.

![Screenshot 2022-12-02 at 10.51.14.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/86347c27-f46b-41b4-9537-834058c38435/Screenshot%202022-12-02%20at%2010.51.14.png)

Avoid controlling checkbox since it leads to strange UX. We should update optimistically and handle errors instead.

Subscribe to `PriceIntent` updates client side so UI is kept up to date.

## Justify why they are needed

I refactored a bit to subscribe to `PriceIntent` updates using Apollo Client. Not sure if this is the direction we want to move in generally. Happy to hear feedback on this!

## Jira issue(s): []

## Checklist before requesting a review

- [ ] I have performed a self-review of my code