Fix schema for price intent update mutation (#1013)
## Describe your changes

Update schema for `PriceIntentDataUpdate` mutation.

Exclude React props from being passed to DOM in `ImageBlock`.

Add `useIsomorphicLayoutEffect` for `useBreakpoint` hook.

## Justify why they are needed

Fix a bunch of warnings in the console.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code