Update ProductPage to be statically rendered (#1034)
## Describe your changes

This is a big one -- make product page statically rendered.

Move management of `PriceIntent` and `ShopSession` client-side.

## Justify why they are needed

We want to be able to show the product page no matter the status of the API.

We want to be able to mix and match product pages anywhere in the URL structure.

## Jira issue(s): []

## Checklist before requesting a review

- [ ] I have performed a self-review of my code