Remove layout from CartPage (#1099)
## Describe your changes

Remove layout + menu from `CartPage`.

![Screenshot 2022-12-15 at 15.42.28.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/49ce5153-51b0-41ad-898d-86f60325baa0/Screenshot%202022-12-15%20at%2015.42.28.png)

Handle closing cart by navigating back to previous route (or store).

## Justify why they are needed

I'm always scared to rely on "history.back" since it doesn't always behave predictably. Like if the history has changes because of "#" updates...

The current solution might be a bit overengineered but I think it's okey.

Probably in the future we should just have a dialog instead of treating it as a separate page.

## Jira issue(s): []

## Checklist before requesting a review

- [ ] I have performed a self-review of my code