Avoid infinite loop of creating price intents after API error (#1137)
## Describe your changes

Avoid causing infinite loop when getting API error after creating new price intents.

## Justify why they are needed

We didn't consider the case where the API would return an error after creating a new price intent. This would cause an infinite loop.

We still have the problem that we run the mutation twice, but that's a separate issue.

## Jira issue(s): []

## Checklist before requesting a review

- [ ] I have performed a self-review of my code