Allow user to edit after seeing price and confirm again (#1158)
## Describe your changes

Fix a bunch of small bugs related to editing a price intent that already is confirmed.

Keep selected product offer in state rather than deriving it from the ID.

Update `ProductOffer` fragment to fetch and use `displayName` for product variants.

Update `CarMileageField` to use placeholder instead of label.

## Justify why they are needed

Bug that didn't allow us to edit a confirmed price intent.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code