Update SE House form (#1166)
## Describe your changes

Update and translate most of SE House form.

![Screenshot 2022-12-23 at 10.53.43.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/f1cac9b7-1bbe-40ea-b7ca-8017e8697736/Screenshot%202022-12-23%20at%2010.53.43.png)

Refactor select to use either label or placeholder.

## Justify why they are needed

I left extra building input update for a different PR.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code