Fix ExtraBuildings field and translate (#1167)
## Describe your changes

Fix issues in the extra buildings field and translate it.

Make sure to not progress to next step after adding/removing an extra building.

## Justify why they are needed

The input still doesn't match design but it's better than before.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code