Update design of extra building field (#1187)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Update design of extra building field.

![Screenshot 2022-12-27 at 16.03.33.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/7a2a25b0-31de-4fa1-bc15-e99d05c817af/Screenshot%202022-12-27%20at%2016.03.33.png)

![Screenshot 2022-12-27 at 16.03.43.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/fb4d5d25-808a-49d4-932c-c4522345b4c9/Screenshot%202022-12-27%20at%2016.03.43.png)

Extract ToggleCard component (switch/checkbox).

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We don't have design for this field yet, but it's a good start.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code