Update Cart Entry Removal dialog (#1189)
## Describe your changes

Create a fullscreen dialog that can be reused for confirmation dialogs and error dialogs.

![Screenshot 2022-12-28 at 13.36.48.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/7900dc9b-4337-46dd-b72f-99ca9da68e47/Screenshot%202022-12-28%20at%2013.36.48.png)

## Justify why they are needed

Emerging as a comopnent in the design system.

Didn't give it specific desktop styles since it's a bit unclear in Figma. Error dialogs are fullscreen even on desktop but not sure how other cases will work.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code