Fix scroll to add to cart button (#1204)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Fix some style issues.

![Screenshot 2022-12-29 at 14.06.29.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/f9afd2f8-02db-4532-a67d-ebc28f2786c5/Screenshot%202022-12-29%20at%2014.06.29.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code