Add new mutation to create internal data collection ID (#1213)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Update Insurely implementation to use new `externalInsuranceProvider.initiateIframeDataCollection` mutation.

It's called as soon as we get the collection ID from Insurely and exchanges it for an internal ID. This is stored in React state until data collection is complete which is when we run the mutation to update the price intent.

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

This is needed for backwards compatability.

Component is getting pretty complex - should refactor soon.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code