Add visual comparison table component (#1216)
## Describe your changes

Add visual `ComparisonTable` to display in purchase form to show the difference between tiers.

![Screenshot 2023-01-02 at 09.41.42.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/87ceac84-f1f7-41ba-8f41-ccb2a26045d5/Screenshot%202023-01-02%20at%2009.41.42.png)

## Justify why they are needed

Just stuck with a basic HTML table. It's only possible to make all columns the same width so that's why I added `min-width` property to the data cells to mimic the design. I think it works well enough.

## Jira issue(s): [GRW-1996]

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code


[GRW-1996]: https://hedvig.atlassian.net/browse/GRW-1996?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ