Enable overflow scroll in offer presenter (#1230)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

See title.

[Screen Recording 2023-01-03 at 10.32.38.mov](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/f32ee03d-55ad-425b-8768-5c3c622a4e59/Screen%20Recording%202023-01-03%20at%2010.32.38.mov)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

When displaying tier selector, the content is overflowing so you can't scroll down to the "add to cart" button.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code