Implement Cart details sheet (#1259)
## Describe your changes

Implement Details Sheet for Cart and Checkout (part of Cart Inventory).

![Screenshot 2023-01-05 at 12.16.07.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/8f4e085a-d962-4849-bcc8-66e47445cc9f/Screenshot%202023-01-05%20at%2012.16.07.png)

![Screenshot 2023-01-05 at 12.16.33.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/78088889-e79c-4f86-af62-5bf29a03b6ce/Screenshot%202023-01-05%20at%2012.16.33.png)

Add documents to Cart query to show in new Details Sheet.

Make changes to `FullscreenDialog` to be more flexible.

![Screenshot 2023-01-05 at 12.17.38.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/00c5cad1-c4da-4e5c-916c-bb2077f1eeed/Screenshot%202023-01-05%20at%2012.17.38.png)

Add extra properties to UI Button to work as a link.

## Justify why they are needed

I'm using some placeholders for the data rows since we don't have it exposed in the API yet.

We should move out also the "confirm remove" dialog to a similar component but we can do that in the next step.

## Jira issue(s): []

## Checklist before requesting a review

- [ ] I have performed a self-review of my code