Refactor Cart + Checkout layout (#1261)
## Describe your changes

Refactor Cart + Checkout page layout to match design.

![Screenshot 2023-01-05 at 14.11.02.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/c6476e71-b089-47bf-b4e2-e86076647420/Screenshot%202023-01-05%20at%2014.11.02.png)

Move padding from shared `CartInventory` to Cart page.

![Screenshot 2023-01-05 at 14.12.31.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/066421fc-c313-45b6-b7e2-dd712cf5fd43/Screenshot%202023-01-05%20at%2014.12.31.png)

## Justify why they are needed

Layout should be decided by the component that implements it, not by the child component. Therefore it's better to use e.g. `Space` in the parent component rather than pushing content with padding in the child component.

In this case, the added padding in `CartEntryList` didn't work on the Checkout page.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code