Show general error when not able to confirm price intent (#1266)
## Describe your changes

Show a general error dialog when we are unable to confirm a price intent.

[Screen Recording 2023-01-05 at 15.51.04.mov](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/a9c4f57b-e3d1-445f-bb2a-ce5210e0abe9/Screen%20Recording%202023-01-05%20at%2015.51.04.mov)
(The text is left aligned but I fixed it after making the recording)

Refactor purchase form a little bit to make this possible.

## Justify why they are needed

I keep adding things to this module -- just to not optimize prematurely. Now we have almost all the features in place so time to rethink it soon I think.

We need to show more specific errors as well but for now this is all we have.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code