Ensure `TextField` takes up the full width of containers (#1286)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

See title.

![Screenshot 2023-01-08 at 18.32.56.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/d9a62c58-a4c0-46eb-b5d2-e646744ee38c/Screenshot%202023-01-08%20at%2018.32.56.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code