Update InputDate design to match Figma (#1296)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Improve the design of the input data field to better match design and other input fields.

![Screenshot 2023-01-09 at 22.25.01.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/193f4efa-006e-4ff9-84e9-c2461be8b1c5/Screenshot%202023-01-09%20at%2022.25.01.png)

![Screenshot 2023-01-09 at 22.25.08.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/b46f3b78-c65a-4820-8c3d-e5648526a463/Screenshot%202023-01-09%20at%2022.25.08.png)

![Screenshot 2023-01-09 at 22.25.17.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/43f6e651-8c8d-454c-8cc4-b266d9e7ccf0/Screenshot%202023-01-09%20at%2022.25.17.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code