Update design of Footer block component (#1332)
## Describe your changes

Update footer block design to match latest Figma design.

![Screenshot 2023-01-13 at 16.05.39.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/3fc502af-d744-4504-8c68-7b09e0525e84/Screenshot%202023-01-13%20at%2016.05.39.png)

![Screenshot 2023-01-13 at 16.05.53.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/38d481a9-af4a-425f-a72f-982dd2c3b95c/Screenshot%202023-01-13%20at%2016.05.53.png)

## Justify why they are needed

We might want to do some refactoring of the Storyblok setup but we can do that later.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code