Refactor to match recommended ESLint setup (#1362)
## Describe your changes

Refactor ESLint setup to match the recommended example from Turborepo: https://github.com/vercel/turbo/tree/main/examples/basic.

Add new workspace `eslint-config-custom` that contains the shared ESLint config. Remove the `scripts` workspace.

Use yarn instead of npx in pre-commit (consistency).

Update imports in `ui` workspace that seems to have been broken since before.

## Justify why they are needed

I was having issues with `lint-staged` not ordering my imports according to our rules. The new setup solves this issue by adding a `.eslintrc.js` config in the root of the project that is used by `lint-staged`.

The recommended way to setup ESLint with Turborepo has also changes since we set it up.

I personally like this setup better since it's easier to extend the shared config compared to before.

There is also a `eslint-config-turbo` but it was causing a lot of errors so let's introduce it separately if at all.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code