Integrate product recommendations on Cart page (#1358)
## Describe your changes

Integrate list of product recommendations on Cart page.

Add recommendations when fetching `ShopSession`.

Refactor Cart page to not use dynamic theme.

![Screenshot 2023-01-18 at 13.20.15.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/fd1cddbe-ff2f-4ee1-9392-8ea4aea69acd/Screenshot%202023-01-18%20at%2013.20.15.png)

![Screenshot 2023-01-18 at 13.19.56.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/9980df20-ae71-43db-b6e0-307e59c27d33/Screenshot%202023-01-18%20at%2013.19.56.png)

## Justify why they are needed

We should probably consider only fetching recommendations on the Cart page and not with every shop session request but don't want to optimize too early.

The API is not 100% working so that's why I didn't include photos.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code