`TextField`: avoid label overflow (#1373)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Add ellipsis overflow to label in `TextField`.

![Screenshot 2023-01-19 at 10.36.51.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/5293e9d2-c95a-4fec-815f-982bb3c958e5/Screenshot%202023-01-19%20at%2010.36.51.png)

Refactor `TextField` to not use dynamic theme.

Use `data-active` prop instead of passing variable to emotion.

## Justify why they are needed

Breaks the design if the label overflows the container.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code