Ensure initiate_gtm is the first event we report (#1422)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Reorder analytics reporting to ensure initiate_gtm is the first event to report

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Prevent missing context when processing events in GTM

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code