Implement Switching Assistant page (#1414)
## Describe your changes

Implement first iteration of Switching Assistant page as part of Checkout.

![Screenshot 2023-01-23 at 18.09.40.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/d9d3db7d-b867-4099-bbff-a32c664a91ca/Screenshot%202023-01-23%20at%2018.09.40.png)

![Screenshot 2023-01-23 at 18.16.15.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/98701cd3-e0f5-4c64-9c51-f476d46657cd/Screenshot%202023-01-23%20at%2018.16.15.png)

## Justify why they are needed

I haven't implemented the "read more" functionality yet.

There's a few changes we need to make to the API to implement all features.

We still need to conditionally navigate to this page if member has any cart entries to cancel.

It still says "confirmation" in the breadcrumbs because I'm mocking the data.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code