Skip downloading en-* locales from Lokalise (#1423)
## Describe your changes

Skip en-* locales when downloading from Lokalise (in "store")

## Justify why they are needed

We rely on falling back to the "en" locale across all countries.

We can easily add them back into the script if we see a need for it.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code