Add API route to redirect to relevant app store (#1426)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Add API route to redirect user to the relevant app store.

## Justify why they are needed

Will be used to generate a QR code that can be shown on confirmation page if signing up on desktop.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code