Track select_item ecommerce event when user visits product page (#1430)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Subj

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Fist step in conversion funnel.  Easier to analyze when we have separate event for product pages instead of relying on virtual_page_view

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code