Report open_price_calculator event (#1436)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Subj.

- Track when editing initial priceIntent (may happen multiple times on close/reopen)
- Ignore editing info for already created offer

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Another step in conversion tracking

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code