Implement 12 column grid on Checkout page (#1459)
## Describe your changes

Implement the 12 column grid layout on checkout page.

[Screen Recording 2023-01-27 at 16.34.39.mov](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/f42710e7-a039-45f6-998e-ffd9642ff1e0/Screen%20Recording%202023-01-27%20at%2016.34.39.mov)

## Justify why they are needed

Better visual consistency

I had to introduce some additional breakpoints that we should confirm with design team before implemeting this on more pages.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code