Remove SsnSeField, should not be used with new setup (#1486)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Remove SsnSeField and its usages

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

We'll have some sort of "section as reusable unit" configration for SSN, so no field-level config is needed.  To be followed up with section config when we know more

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code