Add peril colors (temp) (#1489)
## Describe your changes

Update design of perils block to match figma

Add colors instead of icons

Add a bunch more colors to the theme needed for peril color icons

![Screenshot 2023-01-30 at 13.47.38.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/885a28d0-fb06-40de-9f65-ca1c4bb023f1/Screenshot%202023-01-30%20at%2013.47.38.png)

## Justify why they are needed

We will hardcode these here until we can get time to add them to PCMS

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code