Cart: hide recommendations if list is empty (#1498)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Hide recommendation list if the list of recommendations is empty

## Justify why they are needed

Otherwise you only see the heading which is odd.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code