Fix displaying car details sheet (#1495)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

See title.

![Screenshot 2023-01-31 at 09.51.50.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/bd6de91e-6b52-4074-8ef8-e19e53c6c683/Screenshot%202023-01-31%20at%2009.51.50.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code