Update insurance list in navigation (#1504)
## Describe your changes

Update insurances section in main navigation.

Refactor navigation a little bit better match Figma design.

![Screenshot 2023-01-31 at 15.41.38.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/c6bf4d20-92a8-44a2-8247-1ee4e556ab0f/Screenshot%202023-01-31%20at%2015.41.38.png)

![Screenshot 2023-01-31 at 15.41.51.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/8dc8d9bb-5aae-4e18-80bb-3db6cd456d1f/Screenshot%202023-01-31%20at%2015.41.51.png)

![Screenshot 2023-01-31 at 15.44.33.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/ab886aa1-0291-48ec-b2f1-748fbe82edaf/Screenshot%202023-01-31%20at%2015.44.33.png)

## Justify why they are needed

I focused on getting it to work and behave like design. However, I believe there's a lot of complexity in this part of the code base. Therefore, we should refactor it soon. I think we should define mobile and desktop styles separetly since a lot of complexity is just because we are trying to use the same components for two different layouts.

I excluded the button block list since design only supports a single link that leads to the store page.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code