Make checkout payment step dynamic (#1515)
## Describe your changes

Only redirect users to payment step if they don't already have a payment method.

Update `CheckoutHeader` to show the correct steps in the breadcrumb.

## Justify why they are needed

Needed for returning members.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code