Update confirmation page to fetch CMS content (#1516)
## Describe your changes

Integrate CMS on confirmation page.

Get all content from the CMS.

## Justify why they are needed

This is 90% a content page so the source should be the CMS.

This is a first implementation, we will need to update this to support rearranging the content blocks in the future.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code