Only resume playing videos after closing modal (#1522)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Keep track whether videos were playing when dialog was opened and resume them when dialog is closed.

## Justify why they are needed

Before we resumed videos even if they were not playing when dialog was opened.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code