Implement insurable limits as accordion (#1548)
## Describe your changes

Update insurable limits component to match design.

Implement as accordion that expands when clicked on.

Remove `Statistc` component (unused).

![Screenshot 2023-02-06 at 11.24.26.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/a3a846c4-0b02-43b1-bcc7-eefc8514d243/Screenshot%202023-02-06%20at%2011.24.26.png)

![Screenshot 2023-02-06 at 11.24.43.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/da6a21cc-cff9-4a2e-8a05-2abfb4669647/Screenshot%202023-02-06%20at%2011.24.43.png)

## Justify why they are needed

Haven't animated accordion on desktop yet but close to what we want.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code