Improve returning member authentication UI (#1529)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Incremental improvements for returning member authentication
- Apply design to BankID login prompt
- Localize affected components

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Make returning member flow look good. To be followed by other PRs with the same goal

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code