Add QR code to download app to confirmation page (#1550)
## Describe your changes

Implement download app qr code on confirmation page.

![Screenshot 2023-02-06 at 14.11.04.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/e0dbacf8-b1d1-4b64-9566-8cb437bddc11/Screenshot%202023-02-06%20at%2014.11.04.png)

Fix small syntax issues.

## Justify why they are needed

The QR code points to staging deployment now but we should update that once we have the production deployment.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code