Update gap between perils to match design (#1572)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

See title.

  ![Screenshot 2023-02-08 at 11.15.42.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/0c6f539d-f43a-4a67-8606-f9cfdf5c15d7/Screenshot%202023-02-08%20at%2011.15.42.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code