Extract BankId login from SsnSeSection into app-wide service available through context (#1565)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Create BankIdContext to handle authentication as app-global concern
- Add global BankIdDialog with authentication UI

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

- Ensure BankId auth is always on top of any modal dialogs
- Prepare to handle signing the same way

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code