Refactor signing to use BankIdContext (#1578)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

- Move signing logic to BankIdContext
- Use same progress dialog for login and signing (some texts are different)
- useReducer-based state machine for BankIdContext

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Unify signing and identification logic and UI

## Jira issue(s): [GRW-2199]

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code


[GRW-2199]: https://hedvig.atlassian.net/browse/GRW-2199?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ