Fix padding in non-full-bleed image block (#1590)
## Describe your changes

Add wrapper layout div with padding to image block that isn't fullbleed.

![Screenshot 2023-02-09 at 08.55.03.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/015a75a2-4da2-49a3-b42e-599a221ab721/Screenshot%202023-02-09%20at%2008.55.03.png)

![Screenshot 2023-02-09 at 08.54.38.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/eb8cceed-80d9-4fd1-b1e1-9229cbbf13db/Screenshot%202023-02-09%20at%2008.54.38.png)

## Justify why they are needed

When we added padding to the same div as the relative wrapper, it caused the image to ignore the padding. This fixes that.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code