Expand content on confirmation page (#1597)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Expand content on confirmation page. Keep 6 column layout for xl media query.

Create override layout componentn for confirmation page.

![Screenshot 2023-02-09 at 11.59.01.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/a2ce6806-7d5d-406b-be61-058bd49f4f61/Screenshot%202023-02-09%20at%2011.59.01.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

Let's see if we really need a special layout for this page in the future.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code