Update styles for product doc links (#1599)
<!--
PR title: GRW-123 / Feature / Awesome new thing
-->

## Describe your changes

Updated styles.

![Screenshot 2023-02-09 at 13.59.23.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/OgXegTwxM9IeuXHZyw5h/82649903-86df-4cea-8f6c-b4643d76ba5f/Screenshot%202023-02-09%20at%2013.59.23.png)

<!--
What changes are made?
If there are many changes, a list might be a good format.
If it makes sense, add screenshots and/or screen recordings here.
-->

## Justify why they are needed

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code