Animate PDP tabs with menu (#1610)
## Describe your changes

Animated the tabs on PDP with the main site menu.

Slight refactor of the product page block to use `css` function over `CSSObject`.

Please test it out in the preview app!

## Justify why they are needed

One caveat is that there's a slight bug if you go from mobile<->desktop layout in the same session. Then the tabs stick to the wrong place but this should not be a big concern in practise.

## Jira issue(s): []

<!--
If there is a Jira issue, add the key (e.g. GRW-123) between the brackets.
A link to that issue will automatically be created.
-->

## Checklist before requesting a review

- [ ] I have performed a self-review of my code